# various batch and batchskl optimizations