--- a/internal/batchskl/iterator.go
+++ b/internal/batchskl/iterator.go
@@ -24,11 +24,6 @@ type splice struct {
 	next uint32
 }
 
-func (s *splice) init(prev, next uint32) {
-	s.prev = prev
-	s.next = next
-}
-
 // Iterator is an iterator over the skiplist object. Use Skiplist.NewIterator
 // to construct an iterator. The current state of the iterator can be cloned
 // by simply value copying the struct.