* internal/batchskl: add in-order insertion fast-path
* internal/batchskl: remove splice.init and Skiplist.set{Next,Prev}
* internal/batchskl: manually inline findSpliceForLevel in findSplice
* batchDecodeStr
* remove batchStorage
* internal/batchskl: rework Skiplist key storage

These were motivated by investigation of the CRDB
`BatchApplyBatchRepr_Pebble` benchmark:

```
name                                                                                 old time/op    new time/op    delta
BatchApplyBatchRepr_Pebble/indexed=false/seq=false/valueSize=10/batchSize=10000-16      188µs ± 4%     148µs ± 2%   -21.22%  (p=0.000 n=10+9)
BatchApplyBatchRepr_Pebble/indexed=false/seq=true/valueSize=10/batchSize=10000-16       185µs ± 4%     148µs ± 1%   -19.80%  (p=0.000 n=10+9)
BatchApplyBatchRepr_Pebble/indexed=true/seq=false/valueSize=10/batchSize=10000-16      2.07ms ± 1%    1.76ms ± 3%   -14.82%  (p=0.000 n=8+10)
BatchApplyBatchRepr_Pebble/indexed=true/seq=true/valueSize=10/batchSize=10000-16       1.44ms ± 3%    0.55ms ± 6%   -61.82%  (p=0.000 n=10+10)
BatchApplyBatchRepr_RocksDB/indexed=false/seq=false/valueSize=10/batchSize=10000-16     566µs ± 1%     543µs ± 1%    -4.02%  (p=0.030 n=2+10)
```